# auth-service/openapi.yaml
openapi: 3.1.0
info:
  title: Auth Service Management API
  version: 1.0.0
  description: Management API for the Goaldone Auth Service. Provides endpoints for user, organization, and invitation management. All endpoints require OAuth2 Bearer token authentication.
  contact:
    name: Goaldone Team
    url: https://goaldone.de
  license:
    name: Proprietary
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://auth.goaldone.de
    description: Production server
security:
  - Bearer: []
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from the OAuth2 authorization server
  schemas:
    UserRequest:
      type: object
      title: UserRequest
      description: Request body for creating or updating a user
      properties:
        password:
          type: string
          description: User password (required for creation)
          example: secure-password-123
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
          description: User account status
          example: ACTIVE
        superAdmin:
          type: boolean
          description: Whether the user has system-wide super-admin privileges
          example: false
        emails:
          type: array
          description: List of email addresses associated with the user (at least one required)
          minItems: 1
          items:
            $ref: '#/components/schemas/EmailRequest'
      required:
        - status
        - emails
    EmailRequest:
      type: object
      title: EmailRequest
      description: Email address information
      properties:
        email:
          type: string
          format: email
          description: Email address
          example: user@example.com
        primary:
          type: boolean
          description: Whether this is the primary email for the user
          example: true
        verified:
          type: boolean
          description: Whether the email address has been verified
          example: false
      required:
        - email
    UserResponse:
      type: object
      title: UserResponse
      description: Response containing user information
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
          description: User account status
          example: ACTIVE
        superAdmin:
          type: boolean
          description: Whether the user has system-wide super-admin privileges
          example: false
        emails:
          type: array
          description: List of email addresses associated with the user
          items:
            $ref: '#/components/schemas/EmailResponse'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
    EmailResponse:
      type: object
      title: EmailResponse
      description: User email address information
      properties:
        id:
          type: string
          format: uuid
          description: Unique email record identifier
          example: 550e8400-e29b-41d4-a716-446655440001
        email:
          type: string
          format: email
          description: Email address
          example: user@example.com
        primary:
          type: boolean
          description: Whether this is the primary email for the user
          example: true
        verified:
          type: boolean
          description: Whether the email address has been verified
          example: true
...
    ProblemDetail:
      type: object
      description: HTTP problem detail response (RFC 7807)
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation of the problem
        instance:
          type: string
          description: A URI reference that identifies the specific occurrence
paths:
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user in the system
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieves a user by their UUID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    put:
      tags:
        - Users
      summary: Update user
      description: Updates an existing user's information
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes a user from the system
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/users/search:
    get:
      tags:
        - Users
      summary: Search user by email
      description: Retrieves a user by their primary email address
      operationId: getUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          description: User primary email
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...